requirements.txt
streamlit
pdfplumber
supabase
python-dotenv
transformers
torch
requests
openai
typing-extensions